Update Skeletal Structure
Updated the list of links in the sidebar of derek.ph which currently lists all posts in my site. The plan is to have an initial structure of the topics that I will be talking about, and then there will be a pillar post on each topic. These pillar posts will then have a link to the multiple posts that I will have.